Rename file useSubmitTransactions to useSubmitTransaction
